requirements.txt
Flask
pymongo[srv]
gunicorn
python-dotenv
